relocates common headers
