+++ test/include/common/cmd.h
