+++ test/include/common/init.h
@@ -1,3 +1,6 @@
+#ifndef GUARD_QUAKE_INIT_H
+#define GUARD_QUAKE_INIT_H
+
 /*
 
 Derived from Quake-II
@@ -21,28 +24,16 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 
 */
 
-// source: main.c -- main program
+// source: init.h -- Quake initializer
 
-#include <stdio.h>
+#ifdef GCC
+int Quake_Free(void);
 
-#include "q_shared.h"
-#include "quake.h"
-#include "common.h"
+int Quake_Init(int const argc, const char **argv)
+__attribute__ ((access (read_only, 2), nonnull (2)));
+#else
+int Quake_Free(void);
+int Quake_Init(int const argc, const char **argv);
+#endif
 
-int main (int const argc, const char **argv)
-{
-	int rc;
-	rc = Quake_Init(argc, argv);
-	if (rc != ERR_ENONE) {
-		return rc;
-	}
-
-//	printf("%s\n", com_argv[0]);
-
-	rc = Quake_Free();
-	if (rc != ERR_ENONE) {
-		return rc;
-	}
-
-	return 0;
-}
+#endif
