+++ test/src/common/cmd/make-inc
@@ -18,14 +18,10 @@ INC = -I../../../include
 # headers
 Q_SHARED_H = ../../../include/q_shared.h
 QUAKE_H    = ../../../include/quake.h
-UTIL_H     = ../../../include/util.h
-COM_H      = ../../../include/com.h
-FS_H       = ../../../include/fs.h
+COMMON_H   = ../../../include/common.h
 HEADERS = $(Q_SHARED_H)\
 	  $(QUAKE_H)\
-	  $(UTIL_H)\
-	  $(COM_H)\
-	  $(FS_H)
+	  $(COMMON_H)
 
 # sources
 CMD_C = cmd.c
