+++ test/src/common/com/com.c
@@ -29,7 +29,8 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 #include <stdarg.h>
 
 #include "q_shared.h"
-#include "com.h"
+#include "quake.h"
+#include "common.h"
 
 #define STDC17 GNU_STD17
