+++ test/src/common/com/make-inc
@@ -17,9 +17,11 @@ INC = -I../../../include
 
 # headers
 Q_SHARED_H = ../../../include/q_shared.h
-COM_H      = ../../../include/com.h
+QUAKE_H    = ../../../include/quake.h
+COMMON_H   = ../../../include/common.h
 HEADERS = $(Q_SHARED)\
-	  $(COM_H)
+	  $(QUAKE_H)\
+	  $(COMMON_H)
 
 # sources
 COM_C = com.c
