+++ test/src/common/cvar/make-inc
@@ -18,14 +18,10 @@ INC = -I../../../include
 # headers
 Q_SHARED_H = ../../../include/q_shared.h
 QUAKE_H    = ../../../include/quake.h
-COM_H      = ../../../include/com.h
-UTIL_H     = ../../../include/util.h
-CMD_H      = ../../../include/cmd.h
+COMMON_H   = ../../../include/common.h
 HEADERS = $(Q_SHARED_H)\
 	  $(QUAKE_H)\
-	  $(COM_H)\
-	  $(UTIL_H)\
-	  $(CMD_H)
+	  $(COMMON_H)
 
 # sources
 CVAR_C = cvar.c
