+++ test/src/common/fs/fs.c
@@ -29,12 +29,7 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 
 #include "q_shared.h"
 #include "quake.h"
-#include "com.h"
-#include "util.h"
-#include "cmd.h"
-#include "cbuf.h"
-#include "cvar.h"
-#include "fs.h"
+#include "common.h"
 
 typedef struct {
 	char name[MAX_QPATH];
