+++ test/src/common/fs/make-inc
@@ -18,20 +18,10 @@ INC = -I../../../include
 # headers
 Q_SHARED_H = ../../../include/q_shared.h
 QUAKE_H    = ../../../include/quake.h
-COM_H      = ../../../include/com.h
-UTIL_H     = ../../../include/util.h
-CBUF_H     = ../../../include/cbuf.h
-CMD_H      = ../../../include/cmd.h
-CVAR_H     = ../../../include/cvar.h
-FS_H       = ../../../include/fs.h
+COMMON_H   = ../../../include/common.h
 HEADERS = $(Q_SHARED_H)\
 	  $(QUAKE_H)\
-	  $(COM_H)\
-	  $(UTIL_H)\
-	  $(CBUF_H)\
-	  $(CMD_H)\
-	  $(CVAR_H)\
-	  $(FS_H)
+	  $(COMMON_H)
 
 # sources
 FS_C = fs.c
