+++ test/src/common/util/make-inc
@@ -18,12 +18,10 @@ INC = -I../../../include
 # headers
 Q_SHARED_H = ../../../include/q_shared.h
 QUAKE_H    = ../../../include/quake.h
-COM_H      = ../../../include/com.h
-UTIL_H     = ../../../include/util.h
+COMMON_H   = ../../../include/common.h
 HEADERS = $(Q_SHARED_H)\
 	  $(QUAKE_H)\
-	  $(COM_H)\
-	  $(UTIL_H)
+	  $(COMMON_H)
 
 # sources
 UTIL_C = util.c
